Fixed bugs which occured in last modification. They include:     When observing clocks were not ticking down.     When a games controller changes the information in the tab (title,closeable,etc) wasnt being updated.     When creating a channel tab, previous tells were not being added.
Also added onError methods to Raptor. These can be used when an error occured to notify the user in a standard way.

Changed ThreadService to use only a ScheduleThreadPoolExecutor. Added code to invoke onError in Raptor when an error occurs in a thread. Removed the System.exit when an exception occurs executing the runnable. Instead added a call to show an error message in Raptor.    

git-svn-id: https://raptor-chess-interface.googlecode.com/svn/trunk@140 f0da3330-95b0-11de-b8e1-0bcae535652c