Changed links in a console to be double clickable instead of single clickable. It was annoying when you tried to copy a link, it would always get executed.
Added the ability to minimize a tab folder quadrant and restore it. When a folder is minimized it shows up in the left side with the quadrant number. To restore it click on the quadrant.

Added a top toolbar just to see what it looks like. I might remove this later on I have'nt decided.

git-svn-id: https://raptor-chess-interface.googlecode.com/svn/trunk@211 f0da3330-95b0-11de-b8e1-0bcae535652c