Fixed a huge bug. All the parsing was being done asychronusly for all events.
This explains why games were ending strangely with *.
The last move was not being shown in games if it was checkmate some times.
It probably fixed the game errors, moves were being made out of sequence.
It may have also fixed the lecture bot issues.

I am going to watch it closely, and close them out if i dont see any after using it for a few days.

git-svn-id: https://raptor-chess-interface.googlecode.com/svn/trunk@2108 f0da3330-95b0-11de-b8e1-0bcae535652c